Added script file
This script grabs the transaction ids from a txt file and accepts a date as parameter and then downloads all the log files from the google cloud bucket for that specific date and then in those files searches if the transaction ids grabbed earlier is present or not. The ids present in the file are exported in a separate txt file. 
Visual animations and error handling are done.